+++ lib/td_cache/concept_cache.ex
@@ -37,8 +37,8 @@ defmodule TdCache.ConceptCache do
   @doc """
   Creates cache entries relating to a given concept.
   """
-  def put(concept) do
-    GenServer.call(__MODULE__, {:put, concept})
+  def put(concept, opts \\ []) do
+    GenServer.call(__MODULE__, {:put, concept, opts})
   end
 
   @doc """
@@ -92,8 +92,8 @@ defmodule TdCache.ConceptCache do
   @doc """
   Deletes cache entries relating to a given concept id.
   """
-  def delete(id) do
-    GenServer.call(__MODULE__, {:delete, id})
+  def delete(id, opts \\ []) do
+    GenServer.call(__MODULE__, {:delete, id, opts})
   end
 
   @doc """
@@ -128,8 +128,8 @@ defmodule TdCache.ConceptCache do
   end
 
   @impl true
-  def handle_call({:put, concept}, _from, state) do
-    reply = put_concept(concept)
+  def handle_call({:put, concept, opts}, _from, state) do
+    reply = put_concept(concept, opts)
     {:reply, reply, state}
   end
 
@@ -173,8 +173,8 @@ defmodule TdCache.ConceptCache do
   end
 
   @impl true
-  def handle_call({:delete, id}, _from, state) do
-    reply = delete_concept(id)
+  def handle_call({:delete, id, opts}, _from, state) do
+    reply = delete_concept(id, opts)
     {:reply, reply, state}
   end
 
@@ -248,7 +248,10 @@ defmodule TdCache.ConceptCache do
   defp concept_entry_to_map(nil), do: nil
 
   defp concept_entry_to_map(%{} = concept) do
-    domain = get_domain(Map.get(concept, :domain_id))
+    domain =
+      concept
+      |> Map.get(:domain_id)
+      |> get_domain()
 
     shared_to_ids =
       concept
@@ -267,7 +270,7 @@ defmodule TdCache.ConceptCache do
     end
   end
 
-  defp delete_concept(id) do
+  defp delete_concept(id, opts) do
     commands = [
       ["DEL", "business_concept:#{id}"],
       ["SREM", @keys, "business_concept:#{id}"],
@@ -279,7 +282,7 @@ defmodule TdCache.ConceptCache do
     results = Redix.transaction_pipeline!(commands)
     [_, _, inactivated, _, _] = results
 
-    unless inactivated == 0 do
+    if opts[:publish] != false && inactivated != 0 do
       publish_event("remove_concepts", id)
     end
 
@@ -290,7 +293,7 @@ defmodule TdCache.ConceptCache do
     ["SISMEMBER", @confidential_ids, id] |> Redix.command!()
   end
 
-  defp put_concept(%{id: id} = concept) do
+  defp put_concept(%{id: id} = concept, opts) do
     shared_to_ids =
       concept
       |> Map.get(:shared_to_ids, [])
@@ -315,7 +318,7 @@ defmodule TdCache.ConceptCache do
     results = Redix.transaction_pipeline!(commands)
     [_, _, _, activated, _, _] = results
 
-    unless activated == 0 do
+    if opts[:publish] != false && activated != 0 do
       publish_event("restore_concepts", id)
     end